+++ ZKLog.m
@@ -21,7 +21,7 @@ NSString* const ZKLogToFileKey = @"ZKLogToFile";
 		NSString *label = [self levelToLabel:level];
 		NSString *now = [self.dateFormatter stringFromDate:[NSDate date]];
 		if (label) {
-			NSString *line = [NSString stringWithFormat:@"%@ [%i] %@ %@ (%@:%u)", now, self.pid, label, message, [[NSString stringWithUTF8String:sourceFile] lastPathComponent], lineNumber];
+			NSString *line = [NSString stringWithFormat:@"%@ [%i] %@ %@ (%@:%lu)", now, self.pid, label, message, [[NSString stringWithUTF8String:sourceFile] lastPathComponent], (unsigned long)lineNumber];
 			fprintf(stderr, "%s\n", [line UTF8String]);
 			fflush(stderr);
 		}
